removed cURL support
